+++ 3-Zinyakova-D20/3-ZinyakovaKate-D20/SubsetHeader.h
@@ -12,16 +12,16 @@
 extern "C" {
 #endif
 
-	int FindSet(unsigned long int** subset, int* mass, int** res, int n, int sum);
-	int SubsetSum(int* set, int n, unsigned long int sum, int** res );
+	int FindSet(int** subset, int* mass, int** res, int n, int sum);
+	int SubsetSum(int* set, int n, unsigned long long sum, int** res );
 	int MassInFile(char* stream, int* mass, int size);
-	void PrintRectMatrix(int** subset, int n, unsigned long int sum);
-	void FillZeroRectMatrix(int** matrix, int str, unsigned long int col);
-	int** CreateRectMatrix(int str, unsigned long int col);
+	void PrintRectMatrix(int** subset, int n, unsigned long long sum);
+	void FillZeroRectMatrix(int** matrix, int str, unsigned long long col);
+	int** CreateRectMatrix(int str, unsigned long long col);
 	void PrintMassConvert(int* mass, int size);
 	void PrintMass(int* mass, int size);
-	int ReadData(char* filename, unsigned long int* sum, int* amount, int** mass);
-	void DestroyMatrix(unsigned long int** matrix, int str, unsigned long int col);
+	int ReadData(char* filename, unsigned long long* sum, int* amount, int** mass);
+	void DestroyMatrix(int** matrix, int str, unsigned long long col);
 	int LabSolution(char* input, char* output);
 
 #ifdef __cplusplus
